refactor: 도형 핸들러 커스텀 훅
